refactor: add folders for pages and components and add css files for each component
