+++ src/index.js
@@ -6,14 +6,14 @@ import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { Provider } from "react-redux";
 
-import { Main } from "./pages/Main";
+import { Main } from "./pages/Main/Main";
 import { Signup } from "./pages/Signup/Signup";
 import { Signin } from "./pages/Signin/Signin";
-import { Catalog } from "./pages/Catalog";
-import { Profile } from "./pages/Profile";
+import { Catalog } from "./pages/Catalog/Catalog";
+import { Profile } from "./pages/Profile/Profile";
 import { ProductsPage } from "./pages/ProductsPage";
-import { Cart } from "./pages/Cart";
-import { Favorites } from "./pages/Favorites";
+import { Cart } from "./pages/Cart/Cart";
+import { Favorites } from "./pages/Favorites/Favorites";
 import store from "./redux/store";
 
 import "./index.css";
