+++ src/pages/Cart/Cart.jsx
@@ -1,3 +1,5 @@
+import "./index.css";
+
 export const Cart = () => {
