+++ src/pages/Favorites/Favorites.jsx
@@ -1,3 +1,5 @@
+import "./index.css";
+
 export const Favorites = () => {
 
     return <div>
