+++ src/pages/Main/Main.jsx
@@ -1,4 +1,5 @@
 import React from "react";
+import "./index.css";
 
 
 export const Main = () => {
