+++ src/pages/Products/index.css
@@ -1,0 +1,63 @@
+.product {
+    width: auto;
+    height: 100%;
+    display: grid;
+    box-shadow: 0px 0px 6px #22222275;
+    border-radius: 15px;
+    padding: 2em;
+    box-sizing: border-box;
+    justify-items: center;
+    align-items: flex-start;
+    text-align: center;
+    text-decoration: none;
+}
+
+.product:hover {
+    scale: 1.01;
+    box-shadow: 0px 0px 10px #222222b7;
+    transition-duration: 300ms;
+    cursor: pointer;
+}
+
+.product__img {
+    width: auto;
+    height: 5em;
+    border-radius: 15%;
+    align-self: flex-start;
+}
+
+.product__name {
+    text-decoration: none;
+    font-size: 14px;
+    align-self: flex-start;
+}
+
+.product__wight {
+    text-decoration: none;
+    font-size: 12px;
+    align-self: flex-start;
+    margin: -5px;
+    color: #b16910;
+}
+
+.product__price {
+    text-decoration: none;
+    color: #066f5f;
+    font-weight: 400;
+    font-size: 16px;
+    align-self: flex-end;
+}
+
+.product__btn {
+    border: solid 0px #000000;
+    border-radius: 10px;
+    font-weight: 400;
+    background-color: #fff700;
+    padding: 8px 15px 8px 15px;
+    align-self: flex-end;
+}
+
+.product__btn:hover {
+    text-shadow: 0.3px 0 0 #000000;
+    cursor: pointer;
+}
