+++ src/pages/Profile/index.css
@@ -1,0 +1,47 @@
+.profile {
+    padding: 0px 100px 0px 100px;
+    text-align: center;
+}
+
+.profile__title {
+    font-weight: 400;
+    padding-top: 15px;
+    padding-bottom: 25px;
+    text-align: center;
+}
+
+.profile__welcome-text {
+    font-weight: 400;
+}
+
+.profile__text {
+    font-size: 14px;
+    font-style: oblique;
+}
+
+.profile__avatar {
+    display: inline-block;
+    width: 33%;
+    height: auto;
+}
+
+.profile__avatar img {
+    width: 8em;
+    height: auto;
+    clip-path: circle(50%);
+    object-fit: cover;
+}
+
+.profile__btn {
+    border: solid 0px #000000;
+    border-radius: 10px;
+    color: #000000 ;
+    font-weight: 400;
+    background-color: #fff700;
+    padding: 8px 15px 8px 15px;
+}
+
+.profile__btn:hover {
+    text-shadow: 0.3px 0 0 #000000;
+    cursor: pointer;
+}
